Architecture redesign
Graphics and ImageConverter replaced with Tileset and Image.  Tileset is now
more like a folder (modelled on libgamearchive's Archive class) and Image is
more functional.  FATTileset has been introduced which mirrors FATArchive
from libgamearchive, to handle much of the data manipulation for those many
tileset formats which are a bit like archives.

Tests still need to be redesigned.