Word Rescue: Remove "extra" tileset in favour of the game description XML file
This tileset had a bunch of image coordinates so that a single 320x200 PCX
image could be broken down into individual tiles.  I think this sort of thing
better belongs in the game XML file, or at the very least in the map function
that converts tile codes into images.